requirements.txt
absl-py==1.4.0
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.16.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argilla==1.3.0
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.0
Automat==22.10.0
backoff==2.2.1
black==21.4b2
blobfile==2.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
coloredlogs==15.0.1
comm==0.2.0
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.1
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.33
dataclasses-json==0.5.7
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.13
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d1e04565d3bec8719335b88be9e9b961bf3ec464
distro==1.8.0
Django==4.1.7
dnspython==2.3.0
effdet==0.3.0
elastic-transport==8.10.0
elasticsearch==8.11.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.92.0
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.16.1
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.47.2
grpcio-tools==1.47.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.1
humanfriendly==10.0
hydra-core==1.3.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iopath==0.1.9
ipykernel==6.27.1
ipython==8.18.1
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.19.1
jmespath==1.0.1
joblib==1.2.0
jsons==1.6.3
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.4
langchain==0.0.94
layoutparser==0.3.4
llama-index==0.4.14
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mmh3==3.0.0
monotonic==1.6
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.14.0
openai==1.2.3
opencv-python==4.6.0.66
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
parsel==1.7.0
parso==0.8.3
pathspec==0.11.0
pdf2image==1.16.2
pdfminer.six==20221105
pdfplumber==0.8.0
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.0.0
portalocker==2.7.0
prompt-toolkit==3.0.41
Protego==0.2.1
protobuf==3.20.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
pycocotools==2.0.6
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.5
PyDispatcher==2.0.7
pydot==1.4.2
Pygments==2.17.2
pymilvus==2.2.4
pymongo==4.3.3
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-multipart==0.0.5
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.1.1
qdrant-client==1.1.3
queuelib==1.6.2
regex==2022.10.31
reportlab==4.0.7
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
schedule==1.2.1
scikit-learn==1.3.0
scipy==1.10.1
Scrapy==2.8.0
scrapy-splash==0.9.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.6.3
starlette==0.25.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.2.0
tika==2.6.0
tiktoken==0.2.0
timm==0.6.12
tldextract==3.4.0
tokenizers==0.13.2
toml==0.10.2
torch==1.13.1
torchvision==0.14.1
tornado==6.3.3
tqdm==4.64.1
traitlets==5.14.0
transformers==4.26.1
Twisted==22.10.0
typing-inspect==0.8.0
typing_extensions==4.5.0
typish==1.9.3
ujson==5.4.0
unstructured==0.4.11
unstructured-inference==0.2.7
urllib3==1.26.14
uvicorn==0.20.0
validators==0.19.0
w3lib==2.1.1
Wand==0.6.11
wcwidth==0.2.12
weaviate-client==3.12.0
Werkzeug==2.2.3
wrapt==1.14.1
XlsxWriter==3.0.8
yacs==0.1.8
yarl==1.8.2
zope.interface==5.5.2
